wgengine/filter: split "local IP" lookups by IPv4-vs-IPv6
If we already know it's an incoming IPv4 packet, no need to match
against the set of IPv6s and vice versa.

Compared to HEAD^:

    goos: darwin
    goarch: arm64
    pkg: tailscale.com/wgengine/filter
                        │    after     │                after3                │
                        │    sec/op    │    sec/op     vs base                │
    FilterMatch/file1-8   18.870n ± 1%   9.660n ± 16%  -48.81% (p=0.000 n=10)

Compared to e2c0d69c9c:

    goos: darwin
    goarch: arm64
    pkg: tailscale.com/wgengine/filter
                        │    before    │                after3                │
                        │    sec/op    │    sec/op     vs base                │
    FilterMatch/file1-8   32.600n ± 1%   9.660n ± 16%  -70.37% (p=0.000 n=10)

Note that so far the benchmark is only exercising rejecting a
non-local packet early, so don't read this as firewall processing
overall gets 70% faster. But part of it does. (I'll rename the
benchmarks to be more explicitly named.)

Updates #12486

Change-Id: I70e5024af03893327d26629a994ab2aa9811f4f3
Signed-off-by: Brad Fitzpatrick <bradfitz@tailscale.com>